// src/modules/products/enums/product-callback-data.enum.ts
export enum ProductCallbackData {
  Continue = 'continue_product:',
  RemoveFromFavorite = 'remove_product_from_favorite:',
  AddToFavorite = 'add_product_to_favorite:',
  Comments = 'product_comments:',
  MoreImages = 'more_product_images:',
  PreviousComments = 'previous_product_comments:',
  NextComments = 'next_product_comments:',
  AddComment = 'add_product_comment:',
}